requirements.txt
pybitcoin
pyyaml
jsonrpclib-pelix
